pipeline/requirements.txt
kafka-python-ng>=2.2.2
boto3>=1.34.0
trino>=0.328.0
requests>=2.31.0
